backend/requirements.txt
Flask
Flask-Cors
google-generativeai
google-cloud-storage
werkzeug
ffmpeg-python
python-dotenv
google-auth>=2.0.0
pyjwt
google-cloud-firestore
flask-jwt-extended